test: 🧪 ParallelShift
